Chore: shares externals
